Restrict category access

+++ server/controllers/categoryController.js
@@ -11,7 +11,7 @@ const getCategories = asyncHandler(async (req, res) => {
 const getCategory = asyncHandler(async (req, res) => {
   try {
     const {id} = req.params;
-    const category = await Category.findById(id);
+    const category = await Category.find({_id: id, userId: req.user.id});
     res.status(200).json(category);
   } catch (error) {
     res.status(400).json({error: error.message});
@@ -33,7 +33,7 @@ const createCategory = asyncHandler(async (req, res) => {
 const deleteCategory = asyncHandler(async (req, res) => {
   try {
     const {id} = req.params;
-    const category = await Category.findOneAndDelete({id});
+    const category = await Category.findOneAndDelete({_id: id, userId: req.user.id});
     res.status(200).json(category);
   } catch (error) {
     res.status(400).json({error: error.message});
@@ -44,9 +44,8 @@ const deleteCategory = asyncHandler(async (req, res) => {
 const updateCategory = asyncHandler(async (req, res) => {
   try {
     const {id} = req.params;
-    const category = await Category.findByIdAndUpdate(id, {
-      ...req.body
-    });
+    const category = await Category.findOneAndUpdate({_id: id, userId: req.user.id}, {...req.body});
+    
     res.status(200).json(category);
   } catch (error) {
     res.status(400).json({error: error.message});
